refactor: clean prismic files
