+++ apps/weshipit/pages/_app.tsx
@@ -3,20 +3,18 @@ import Head from 'next/head';
 import './styles.css';
 
 import { PrismicProvider } from '@prismicio/react';
-import { client } from '../prismic';
+import { client } from './api/prismic';
 
 function CustomApp({ Component, pageProps }: AppProps) {
   return (
-    <>
-      <PrismicProvider client={client}>
-        <Head>
-          <title>weshipit.today — React Native Experts</title>
-        </Head>
-        <main>
-          <Component {...pageProps} />
-        </main>
-      </PrismicProvider>
-    </>
+    <PrismicProvider client={client}>
+      <Head>
+        <title>weshipit.today — React Native Experts</title>
+      </Head>
+      <main>
+        <Component {...pageProps} />
+      </main>
+    </PrismicProvider>
   );
 }
