+++ libs/ui/src/lib/header/header.stories.tsx
@@ -6,7 +6,7 @@ export default {
   title: 'Header',
 } as ComponentMeta<typeof Header>;
 
-const Template: ComponentStory<typeof Header> = (args) => <Header {...args} />;
+const Template: ComponentStory<typeof Header> = () => <Header />;
 
 export const Primary = Template.bind({});
 Primary.args = {};
